requirements.txt
django==2.0.6
django-configurations==2.0
django-extensions==2.0.7
django-debug-toolbar==1.9.1
Pillow==5.1.0
psycopg2-binary==2.7.5
factory-boy==2.11.1
djangorestframework==3.8.2
django-cors-headers==2.4.0
django-nose==1.4.5
coverage==4.5.1
celery==4.2.1
django-celery-results==1.0.1
raven=6.9.0
pytest-django==3.4.3
pytest-sugar==0.9.1
pytest-cov==2.6.0
